Constraints

SeaPearl.Absolute — Type

Absolute(x::SeaPearl.AbstractIntVar, y::SeaPearl.AbstractIntVar, ::SeaPearl.Trailer)

Absolute value constraint, enforcing y = |x|.

SeaPearl.AllDifferent — Type

AllDifferent(x::Array{<:AbstractIntVar}, ::SeaPearl.Trailer)

AllDifferent constraint, enforcing ∀ i ≠ j ∈ ⟦1, length(x)⟧, x[i] ≠ x[j].

The implementation of this contraint is inspired by: https://www.researchgate.net/publication/200034395AFilteringAlgorithmforConstraintsofDifferencein_CSPs Many of the functions below relate to algorithms depicted in the paper, and their documentation refer to parts of the overall algorithm.

SeaPearl.BinaryEquivalence — Type

BinaryEquivalence(x::BoolVar, y::BoolVar, trailer::SeaPearl.Trailer)

Binary equivalence constraint, states that x <=> y.

SeaPearl.BinaryImplication — Type

BinaryImplication(x::BoolVar, y::BoolVar, trailer::SeaPearl.Trailer)

Binary implication constraint, states that x => y.

SeaPearl.BinaryMaximumBC — Type

BinaryMaximum(x::AbstractIntVar, y::AbstractIntVar, z::AbstractIntVar, trailer::SeaPearl.Trailer)

BinaryMaximum constraint, states that x == max(y, z)

SeaPearl.BinaryOr — Type

BinaryOr(x::BoolVar, y::BoolVar, trailer::SeaPearl.Trailer)

Binary Or constraint, states that x || y.

SeaPearl.BinaryXor — Type

BinaryXor(x::AbstractBoolVar, y::AbstractBoolVar, trailer::SeaPearl.Trailer)

Binary Xor constraint, states that x ⊻ y.

SeaPearl.TableConstraint — Type

TableConstraint

Efficient implentation of the constraint enforcing: ∃ j ∈ ⟦1,m⟧, such that ∀ i ∈ ⟦1,n⟧ xᵢ=table[i,j].

The datastructure and the functions using it are inspired by: Demeulenaere J. et al. (2016) Compact-Table: Efficiently Filtering Table Constraints with Reversible Sparse Bit-Sets. In: Rueher M. (eds) Principles and Practice of Constraint Programming. CP 2016. Lecture Notes in Computer Science, vol 9892. Springer, Cham. https://doi.org/10.1007/978-3-319-44953-1_14

TableConstraint(scope, active, table, currentTable, modifiedVariables, unfixedVariables, supports, residues)

TableConstraint default constructor.

Arguments

scope::Vector{<:AbstractIntVar}: the ordered variables present in the table.

table::Matrix{Int}: the original table describing the constraint (impossible assignment are filtered).

currentTable::RSparseBitSet{UInt64}: the reversible representation of the table.

modifiedVariables::Vector{Int}: vector with the indexes of the variables modified since the last propagation.

unfixedVariables::Vector{Int}: vector with the indexes of the variables which are not binding.

supports::Dict{Pair{Int,Int},BitVector}: dictionnary which, for each pair (variable => value), gives the support of this pair.

residues::Dict{Pair{Int,Int},Int}: dictionnary which, for each pair (variable => value), gives the residue of this pair.

TableConstraint(variables, table, supports, trailer)

Create a CompactTable constraint from the variables, with values given in table and supports given in supports.

This constructor gives full control on both table and supports. The attributes are not duplicated and remains linked to the variables given to the constructor. It should be used only to avoid duplication of table when using the same table constraint many times with different variables. WARNING: all variables must have the same domain; table and supports must be cleaned.

Arguments

variables::Vector{<:AbstractIntVar}: vector of variables of size (n, ).

table::Matrix{Int}: matrix of the constraint of size (n, m).

supports::Dict{Pair{Int,Int},BitVector}: for each assignment, a list of the tuples supporting this assignment.

trailer::Trailer: the trailer of the model.