Reinforcement Learning

Neural Networks

SeaPearl.CPNN — Type

CPNN(;
    graphChain::Flux.Chain = Flux.Chain()
    nodeChain::Flux.Chain = Flux.Chain()
    globalChain::Flux.Chain = Flux.Chain()
    outputChain::Union{Flux.Dense, Flux.Chain} = Flux.Chain()
)

This structure is here to provide a flexible way to create a nn model which respect this approach: Making modification on the graph, then extract one node feature and modify it. The CPNN pipeline uses the features of the variable branched on (variableFeatures) and, if specified, the global features of the graph (globalFeatures). Contrary to FullFeaturedCPNN, it does not use the features of the possible values to make predictions. This pipeline is made of 4 networks:

graphChain: a graph convolutional network (GCN). It takes the original featured graph (states.fg) as an input,

nodeChain: a fully connected neural network (FCNN). It takes the features of the variable to branch on as an input,

globalChain: FCNN. It takes the global features of the graph as an input,

outputChain: FCNN. It takes the concatenation of the outputs of nodeChain and globalChain as an input.

SeaPearl.FullFeaturedCPNN — Type

FullFeaturedCPNN(;
    graphChain::Flux.Chain
    nodeChain::Flux.Chain
    globalChain::Flux.Chain
    outputChain::Flux.Dense
)

This structure is here to provide a flexible way to create a nn model which respect this approach: Making modification on the graph, then extract one node feature and modify it.

This pipeline is made of 4 networks:

graphChain: a graph convolutional network (GCN). It takes the original featured graph (states.fg) as an input,

nodeChain: a fully connected neural network (FCNN). It is used both on variableFeatures and valueFeatures.

globalChain: FCNN. It takes the global features of the graph as an input,

outputChain: FCNN. It takes the concatenation of the outputs of nodeChain and globalChain as an input.

Like the CPNN pipeline, the FullFeaturedCPNN pipeline uses the features (variableFeatures) of the variable branched on and, if specified, the global features of the graph (globalFeatures). But contrary to CPNN, it also uses the features (valueFeatures) of the possible values that the variable can be assigned to.

FullFeaturedCPNN generates a Q-table containing the Q-value of all the values, both the possible and impossible ones. A mask then selects the possible values.

SeaPearl.HeterogeneousCPNN — Type

HeterogeneousCPNN(;
    graphChain::Flux.Chain = Flux.Chain()
    nodeChain::Flux.Chain = Flux.Chain()
    globalChain::Flux.Chain = Flux.Chain()
    outputChain::Union{Flux.Dense, Flux.Chain} = Flux.Chain()
)

This structure is here to provide a flexible way to create a nn model which respect this approach: Making modification on the graph, then extract one node feature and modify it.

SeaPearl.HeterogeneousFullFeaturedCPNN — Type

HeterogeneousFullFeaturedCPNN(;
    graphChain::Flux.Chain
    varChain::Flux.Chain = Flux.Chain()
    valChain::Flux.Chain = Flux.Chain()
    globalChain::Flux.Chain
    outputChain::Flux.Dense
)

This structure is here to provide a flexible way to create a nn model which respect this approach: Making modification on the graph, then extract one node feature and modify it. This CPNN is designed to process HeterogeneousFeaturedGraphs.

This pipeline works in the following way :

We apply a GNN on the input featured graph.

We extract the contextualized-feature of the branching variable and pass it trought varChain

We extract the contextualized-feature of each value that is part of the variable's domain of definition and pass it trought valChain.

We concat the two reshaped previous results (optionnaly with a global feature) and pass it trought outputChain to generate the output Q-vector.

SeaPearl.HeterogeneousVariableOutputCPNN — Type

HeterogeneousVariableOutputCPNN(;
    graphChain::Flux.Chain
    nodeChain::Flux.Chain
    outputChain::Flux.Dense
)

This structure is here to provide a flexible way to create a nn model which respect this approach: Making modification on the graph, then extract one node feature and modify it.