Constraints

SeaPearl.Absolute — Type

Absolute(x::SeaPearl.AbstractIntVar, y::SeaPearl.AbstractIntVar, ::SeaPearl.Trailer)

Absolute value constraint, enforcing y = |x|.

Missing docstring.

Missing docstring for SeaPearl.AllDifferent(x::Array{<:SeaPearl.AbstractIntVar}, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.BinaryEquivalence(x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.BinaryImplication(x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.BinaryMaximumBC(x::SeaPearl.AbstractIntVar, y::SeaPearl.AbstractIntVar, z::SeaPearl.AbstractIntVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.BinaryOr(x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.BinaryXor(x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.TableConstraint(variables::Vector{<:SeaPearl.AbstractIntVar}, table::Matrix{Int}, supports::Dict{Pair{Int,Int},BitVector}, trailer::SeaPearl.SeaPearl.Trailer). Check Documenter's build log for details.

SeaPearl.TableConstraint — Method

TableConstraint(variables::Vector{<:AbstractIntVar}, table::Matrix{Int}, trailer::SeaPearl.Trailer)

Create a CompactTable constraint from the variables, with values given in table.

This is the recommended constructor, as it safely builds the constraint and isolates it. If you choose any of the 2 other constructor, beware how you create your attributes, and be sure not to modify them at any time.

Arguments

variables::Vector{<:AbstractIntVar}: vector of variables of size (n, ).

table::Matrix{Int}: matrix of the constraint of size (n, m).

trailer::Trailer: the trailer of the model.

Missing docstring.

Missing docstring for SeaPearl.BinaryImplication(x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Disjunctive(earliestStartingTime::Array{<:SeaPearl.AbstractIntVar}, processingTime::Array{Int}, trailer, filteringAlgorithm::Array{filteringAlgorithmTypes} = [algoTimeTabling]). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.Element1D(matrix::Array{Int, 1}, i::SeaPearl.AbstractIntVar, x::SeaPearl.AbstractIntVar, trailer::SeaPearl.Trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.Element2D(matrix::Array{Int, 2}, x::SeaPearl.AbstractIntVar, y::SeaPearl.AbstractIntVar, z::SeaPearl.AbstractIntVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.EqualConstant(x::SeaPearl.SeaPearl.AbstractIntVar, v::Int, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.GreaterOrEqualConstant(x::SeaPearl.SeaPearl.AbstractIntVar, v::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.InSet(x::SeaPearl.AbstractIntVar, s::SeaPearl.IntSetVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.IntervalConstant(x::SeaPearl.IntVar, lower::Int, upper::Int, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.isBinaryAnd(b::SeaPearl.AbstractBoolVar, x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.isBinaryOr(b::SeaPearl.AbstractBoolVar, x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.isBinaryXor(b::SeaPearl.AbstractBoolVar, x::SeaPearl.AbstractBoolVar, y::SeaPearl.AbstractBoolVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.isLessOrEqual(b::SeaPearl.AbstractBoolVar, x::SeaPearl.AbstractIntVar, y::SeaPearl.AbstractIntVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.LessOrEqualConstant(x::SeaPearl.AbstractIntVar, v::Int, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.LessOrEqual(x::SeaPearl.AbstractIntVar, y::SeaPearl.AbstractIntVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.MaximumConstraint(x::Array{<:SeaPearl.AbstractIntVar},y::SeaPearl.AbstractIntVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.NotEqualConstant(x::SeaPearl.AbstractIntVar, v::Int, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.NotEqual(x::SeaPearl.SeaPearl.AbstractIntVar, y::SeaPearl.SeaPearl.AbstractIntVar, trailer::SeaPearl.Trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.ReifiedInSet(x::SeaPearl.AbstractIntVar, s::SeaPearl.IntSetVar, b::SeaPearl.AbstractBoolVar, trailer::SeaPearl.Trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.SetDiffSingleton(a::SeaPearl.IntSetVar, b::SeaPearl.IntSetVar, x::SeaPearl.AbstractIntVar, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.SetEqualConstant(s::SeaPearl.IntSetVar, c::Set{Int}, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.SumGreaterThan(x::Array{<:SeaPearl.AbstractIntVar}, lower::Int, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.SumLessThan(x::Array{<:SeaPearl.AbstractIntVar}, upper::Int, trailer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeaPearl.SumToZero(x::Array{<:SeaPearl.AbstractIntVar}, trailer). Check Documenter's build log for details.